some reorganization
