+++ src/app/components/Search/Search.module.css
@@ -1,0 +1,34 @@
+.form {
+  max-width: 800px;
+  margin: 0 auto;
+  position: relative;
+  --button-width: 80px;
+}
+
+.input {
+  position: absolute;
+  width: 100%;
+  height: 3rem;
+  border-radius: 39px;
+  border: 0;
+  padding-left: 1rem;
+  font-size: 1rem;
+
+  padding-right: calc(var(--button-width) + 1rem);
+}
+
+.button {
+  position: absolute;
+  right: 0;
+  background: var(--accent-color);
+  width: var(--button-width);
+  height: 3rem;
+  border: 0;
+  border-radius:0 var(--border-radius) var(--border-radius) 0;
+}
+
+.button img {
+  /* Manually center, since the rounded borders make it look off */
+  margin-left: -10px;
+}
+
