+++ src/app/layout.tsx
@@ -1,3 +1,6 @@
+import Head from "next/head";
+import "./styles/globals.css";
+
 export default function RootLayout({
   // Layouts must accept a children prop.
   // This will be populated with nested layouts or pages
@@ -7,6 +10,10 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
+      <Head>
+        <title>Kanjikka - Lookup Kanji Stroke Order</title>
+      </Head>
+
       <body>{children}</body>
     </html>
   );
