+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Home from "@/index";
+import Home from "@/app/components/Home/Home";
 
 export default function Page() {
   return <Home />;
