+++ src/app/styles/globals.css
@@ -14,3 +14,17 @@ a {
 * {
   box-sizing: border-box;
 }
+
+:root {
+  --text-main: #000;
+  --text-alternate: #fff;
+  --accent-color: #CE3B2B;
+
+  --border-radius: 39px;
+}
+
+
+.container {
+  max-width: 1280px;
+  margin: 0 auto;
+}
