--- styles/Home.module.css
@@ -1,15 +1,0 @@
-.container {
-  max-width: 980px;
-  margin: 0 auto;
-}
-
-.title {
-  margin: 0;
-  line-height: 1.15;
-  font-size: 4rem;
-}
-
-.title,
-.description {
-  text-align: center;
-}
